add ElasticCachingDecorator - fix RepositoriesServiceProvider.php - add AbstractCachingDecorator
